app/requirements.txt
Flask==3.1.3
psycopg2-binary==2.9.11
redis==7.2.1
